Add performance tests fo edge detection
